made N a integer since steps are an integer

+++ harmonic_oscillator/diffeq.h
@@ -26,7 +26,7 @@ void euler(Eigen::Vector3d r0,
 void RungeKutta2(Eigen::Vector3d r0,
            Eigen::Vector3d v0,
            double const step,
-           double const N,
+           int const N,
            double const m,
            Eigen::Vector3d F(Eigen::Vector3d));
